refactor: :art: split homepage into components
