+++ client/src/__fixtures__/reduxStore.ts
@@ -50,6 +50,11 @@ const initialMockSearchReults: SearchResultsState = {
     language: [],
     sort: "date",
     type: "all",
+  },
+  inputFocused: {
+    search: false,
+    location: false,
+    theme: false,
   }
 }
