+++ client/src/__tests__/recherche.test.tsx
@@ -57,6 +57,11 @@ describe("recherche", () => {
               language: [],
               sort: "date",
               type: "all"
+            },
+            inputFocused: {
+              search: false,
+              location: false,
+              theme: false
             }
           }
         }
