--- client/src/components/Pages/homepage/MobileAppSection/index.ts
@@ -1,3 +1,0 @@
-import MobileAppSection from "./MobileAppSection";
-
-export default MobileAppSection;
