+++ client/src/components/Pages/homepage/Sections/AllThemes/index.ts
@@ -1,0 +1,3 @@
+import AllThemes from "./AllThemes";
+
+export default AllThemes;
