+++ client/src/components/Pages/homepage/Sections/Community/index.ts
@@ -1,0 +1,3 @@
+import Community from "./Community";
+
+export default Community;
