+++ client/src/components/Pages/homepage/Sections/FreeResources/FreeResources.module.scss
@@ -1,0 +1,4 @@
+.resources_image {
+  max-width: 100%;
+  height: auto;
+}
