+++ client/src/components/Pages/homepage/Sections/FreeResources/index.ts
@@ -1,0 +1,3 @@
+import FreeResources from "./FreeResources";
+
+export default FreeResources;
