+++ client/src/components/Pages/homepage/Sections/HelpUs/index.ts
@@ -1,0 +1,3 @@
+import HelpUs from "./HelpUs";
+
+export default HelpUs;
