+++ client/src/components/Pages/homepage/Sections/Hero/index.ts
@@ -1,0 +1,3 @@
+import Hero from "./Hero";
+
+export default Hero;
