+++ client/src/components/Pages/homepage/Sections/Infos/index.ts
@@ -1,0 +1,3 @@
+import Infos from "./Infos";
+
+export default Infos;
