+++ client/src/components/Pages/homepage/Sections/MainFigures/index.ts
@@ -1,0 +1,3 @@
+import MainFigures from "./MainFigures";
+
+export default MainFigures;
