+++ client/src/components/Pages/homepage/Sections/MobileApp/MobileAppIllu/index.ts
