+++ client/src/components/Pages/homepage/Sections/MobileApp/index.ts
@@ -1,0 +1,3 @@
+import MobileApp from "./MobileApp";
+
+export default MobileApp;
