+++ client/src/components/Pages/homepage/Sections/NewContent/NewContent.module.scss
@@ -1,0 +1,25 @@
+@import "src/scss/mixins/responsive";
+@import "src/scss/mixins/unit";
+@import "src/scss/mixins/text";
+
+.sliders {
+  :global(.content-card-slider) {
+    --gradient-start: #f6f6f600;
+    --gradient-end: #f6f6f6;
+  }
+
+  .title {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    @include isTablet() {
+      & > button {
+        display: none;
+      }
+      h2 {
+        @include text("h5");
+      }
+    }
+  }
+}
