+++ client/src/components/Pages/homepage/Sections/NewContent/index.ts
@@ -1,0 +1,3 @@
+import NewContent from "./NewContent";
+
+export default NewContent;
