+++ client/src/components/Pages/homepage/Sections/WhyAccordions/index.ts
@@ -1,0 +1,3 @@
+import WhyAccordions from "./WhyAccordions";
+
+export default WhyAccordions;
