+++ client/src/components/Pages/homepage/Sections/index.ts
@@ -1,0 +1,10 @@
+export { default as Hero } from "./Hero";
+export { default as AllThemes } from "./AllThemes";
+export { default as MobileApp } from "./MobileApp";
+export { default as NewContent } from "./NewContent";
+export { default as WhyAccordions } from "./WhyAccordions";
+export { default as FreeResources } from "./FreeResources";
+export { default as HelpUs } from "./HelpUs";
+export { default as MainFigures } from "./MainFigures";
+export { default as Community } from "./Community";
+export { default as Infos } from "./Infos";
