--- client/src/scss/pages/homepage.module.scss
@@ -1,146 +1,0 @@
-@import "src/scss/mixins/responsive";
-@import "src/scss/mixins/unit";
-@import "src/scss/mixins/text";
-@import "src/scss/colors";
-
-.container {
-  width: 100%;
-}
-
-// Hero
-.hero {
-  min-height: 750px;
-  background-size: cover;
-  background-repeat: no-repeat;
-  padding-top: u(20);
-
-  @include isTablet() {
-    padding-top: u(10);
-    min-height: 90vh;
-  }
-
-  h1 {
-    text-align: center;
-
-    @include isTablet() {
-      @include text("h4");
-    }
-  }
-
-  @include media-min("tablet-up") {
-    background-image: url(../../assets/figma/integration_illu.svg);
-  }
-  @include isMobile() {
-    background-image: url(../../assets/figma/home_mobile.svg);
-  }
-}
-
-// themes
-.themes {
-  h2 {
-    @include isTablet() {
-      @include text("h4");
-    }
-  }
-}
-
-// Sliders
-.content_sliders {
-  :global(.content-card-slider) {
-    --gradient-start: #f6f6f600;
-    --gradient-end: #f6f6f6;
-  }
-
-  .title_line {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    @include isTablet() {
-      & > button {
-        display: none;
-      }
-      h2 {
-        @include text("h5");
-      }
-    }
-  }
-}
-
-// Resources
-.resources_image {
-  max-width: 100%;
-  height: auto;
-}
-
-// Help
-.help_text {
-  margin-top: u(6);
-  margin-bottom: 0;
-
-  @include isTablet() {
-    margin-top: u(4);
-  }
-}
-
-// Community
-.row {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-
-  @include isTablet() {
-    flex-wrap: nowrap;
-    overflow-x: scroll;
-    justify-content: flex-start;
-    margin-left: -15px;
-    margin-right: -15px;
-    padding-bottom: u(4);
-  }
-}
-
-// Infos
-.infos {
-  overflow-x: hidden;
-}
-.infos_col {
-  max-width: 430px;
-  margin: auto;
-  padding: u(20) 0;
-
-  @include isTablet() {
-    padding: u(10) 0;
-  }
-
-  .action {
-    margin-top: u(10);
-
-    @include isTablet() {
-      margin-top: u(6);
-    }
-  }
-  .img {
-    max-width: 100%;
-  }
-}
-
-.newsletter {
-  position: relative;
-}
-.newsletter_input {
-  input {
-    height: 64px;
-  }
-  & > span {
-    top: 20px !important;
-  }
-  :global(.input-error-icon) {
-    right: 150px !important;
-  }
-}
-.newsletter_submit {
-  position: absolute;
-  right: 8px;
-  top: 8px;
-  height: 48px;
-}
